// types/schemas.ts
export const trackOrderValidationSchema: ZodType<StoreTrackOrderData> =
  z.object({
    _id: z.string().min(20).max(30),
    user_id: z.string().optional().nullable(),
  });